+++ Tests/GauntletLegacyTests/Extensions/DispatchQueueExtensionsTests.swift
@@ -22,7 +22,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 //  SOFTWARE.
 
-import Gauntlet
+import GauntletLegacy
 import Foundation
 import XCTest